a ggplot2 theme which removes labels, ticks and titles from both axes.

a ggplot2 theme which removes labels, ticks and titles from both axes.

Usage

theme_map()

See also

Examples

  library(ggplot2)
  p <- ggplot(mtcars, aes(x = mpg, y = drat)) + geom_point()
  p.map <- p + theme_map()