The theme in compliance with the Flanders style guide version >= 2015.

The theme in compliance with the Flanders style guide version >= 2015.

Usage

theme_vlaanderen2015(base_size = 12, base_family = "", transparent = FALSE)

Arguments

base_size

base font size, given in pts.

base_family

base font family

transparent

Make backgrounds transparent. FALSE: all backgrounds are white, TRUE: all backgrounds are transparent. You can pass a vector to transparent. In that case, it will check whether the values "plot", "panel" and/or "legend" are present. The according items will be transparent. Transparent panel will use grey instead of white grid lines.

See also

Author

Thierry Onkelinx, Nicole De Groof

Examples

  library(ggplot2)
  p <- ggplot(mtcars, aes(x = mpg, y = drat)) + geom_point()
  p.vl <- p + theme_vlaanderen2015()